backend/requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
Markdown==3.8
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
sqlparse==0.5.3
pre-commit==4.2.0
